# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Jewelry Invoice Bot API
  description: |
    API REST para el sistema de facturación de joyerías.

    ## Características

    - **Multi-tenant**: Soporte para múltiples organizaciones
    - **Autenticación**: Via API Key o JWT
    - **Rate Limiting**: Límites por plan de suscripción
    - **Validación**: Validación estricta de datos de entrada

    ## Autenticación

    La API utiliza el header `X-Organization-ID` para identificar la organización.
    En producción, también se requiere autenticación via API Key o JWT.

    ```
    X-Organization-ID: org-xyz-123
    Authorization: Bearer <token>
    ```

    ## Rate Limits

    | Plan       | Requests/min | Facturas/mes |
    |------------|--------------|--------------|
    | Basic      | 60           | 100          |
    | Pro        | 300          | 500          |
    | Enterprise | 1000         | Ilimitado    |

    ## Errores

    La API retorna errores en formato JSON estándar:

    ```json
    {
      "error": "ValidationError",
      "message": "Descripción del error",
      "status_code": 400
    }
    ```

  version: 1.0.0
  contact:
    name: Soporte Técnico
    email: soporte@joyeriainvoice.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Servidor de desarrollo
  - url: https://api.joyeriainvoice.com
    description: Servidor de producción
...
paths:
...
  /api/v1/organizations:
    get:
      tags:
        - organizations
      summary: Listar organizaciones
      description: Retorna lista paginada de organizaciones
      operationId: listOrganizations
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
        - name: include_inactive
          in: query
          description: Incluir organizaciones inactivas
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Lista de organizaciones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrganizationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - organizations
      summary: Crear organización
      description: Crea una nueva organización (tenant)
      operationId: createOrganization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationCreate'
      responses:
        '201':
          description: Organización creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/invoices:
    get:
      tags:
        - invoices
      summary: Listar facturas
      description: Retorna lista paginada de facturas de la organización
      operationId: listInvoices
      parameters:
        - $ref: '#/components/parameters/OrgIdHeader'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
        - name: status
          in: query
          description: Filtrar por estado
          schema:
            $ref: '#/components/schemas/InvoiceStatus'
      responses:
        '200':
          description: Lista de facturas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InvoiceListItem'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - invoices
      summary: Crear factura
      description: Crea una nueva factura
      operationId: createInvoice
      parameters:
        - $ref: '#/components/parameters/OrgIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceCreate'
      responses:
        '201':
          description: Factura creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
...
components:
  parameters:
    OrgIdHeader:
      name: X-Organization-ID
      in: header
      required: true
      description: ID de la organización
      schema:
        type: string
        example: "org-xyz-123"
...
    LimitParam:
      name: limit
      in: query
      description: Número máximo de resultados
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
    OffsetParam:
      name: offset
      in: query
      description: Número de resultados a saltar
      schema:
        type: integer
        minimum: 0
        default: 0
  schemas:
...
    InvoiceStatus:
      type: string
      enum:
        - BORRADOR
        - PENDIENTE
        - PAGADA
        - ANULADA
    PlanType:
      type: string
      enum:
        - basic
        - pro
        - enterprise
    InvoiceItem:
      type: object
      required:
        - descripcion
        - cantidad
        - precio_unitario
      properties:
        descripcion:
          type: string
          minLength: 2
          maxLength: 200
          example: "Anillo Oro 18K"
        cantidad:
          type: integer
          minimum: 1
          maximum: 9999
          example: 1
        precio_unitario:
          type: number
          minimum: 0
          example: 500000
        subtotal:
          type: number
          example: 500000
    InvoiceCreate:
      type: object
      required:
        - cliente_nombre
        - items
      properties:
        cliente_nombre:
          type: string
          minLength: 3
          maxLength: 100
          example: "Juan Pérez"
        cliente_cedula:
          type: string
          minLength: 6
          maxLength: 12
          example: "12345678"
        cliente_telefono:
          type: string
          maxLength: 20
          example: "3001234567"
        cliente_direccion:
          type: string
          maxLength: 200
        cliente_email:
          type: string
          format: email
        items:
          type: array
          minItems: 1
          maxItems: 50
          items:
            $ref: '#/components/schemas/InvoiceItem'
        notas:
          type: string
          maxLength: 500
...
    InvoiceListItem:
      type: object
      properties:
        id:
          type: string
        numero_factura:
          type: string
        cliente_nombre:
          type: string
        cliente_cedula:
          type: string
        subtotal:
          type: number
        impuestos:
          type: number
        total:
          type: number
        estado:
          $ref: '#/components/schemas/InvoiceStatus'
        created_at:
          type: string
          format: date-time
    InvoiceResponse:
      allOf:
        - $ref: '#/components/schemas/InvoiceListItem'
        - type: object
          properties:
            organization_id:
              type: string
            cliente_telefono:
              type: string
            cliente_direccion:
              type: string
            cliente_email:
              type: string
            items:
              type: array
              items:
                $ref: '#/components/schemas/InvoiceItem'
            descuento:
              type: number
            vendedor_id:
              type: integer
            vendedor_nombre:
              type: string
            notas:
              type: string
            updated_at:
              type: string
              format: date-time
    OrganizationCreate:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          example: "Joyería El Diamante"
        plan:
          $ref: '#/components/schemas/PlanType'
        invoice_prefix:
          type: string
          maxLength: 10
          default: "FAC"
          example: "JOY"
        owner_email:
          type: string
          format: email
...
    OrganizationResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        plan:
          $ref: '#/components/schemas/PlanType'
        invoice_prefix:
          type: string
        is_active:
          type: boolean
        users_count:
          type: integer
        invoices_count:
          type: integer
        created_at:
          type: string
          format: date-time
        plan_limits:
          $ref: '#/components/schemas/PlanLimits'
    PlanLimits:
      type: object
      properties:
        invoices_per_month:
          type: integer
        users_per_org:
          type: integer
        max_items_per_invoice:
          type: integer
        features:
          type: object
          properties:
            ai_extraction:
              type: boolean
            voice_input:
              type: boolean
            photo_input:
              type: boolean
            custom_templates:
              type: boolean
            api_access:
              type: boolean
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
        - status_code
      properties:
        error:
          type: string
          example: "NotFound"
        message:
          type: string
          example: "Recurso no encontrado"
        status_code:
          type: integer
          example: 404
        timestamp:
          type: string
          format: date-time
        details:
          type: object
  responses:
    BadRequest:
      description: Datos de entrada inválidos
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "ValidationError"
            message: "El campo 'cliente_nombre' es requerido"
            status_code: 400
    Unauthorized:
      description: No autorizado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Unauthorized"
            message: "Token de autenticación requerido"
            status_code: 401
...
    RateLimited:
      description: Rate limit excedido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "RateLimitExceeded"
            message: "Has excedido el límite de requests. Intenta de nuevo en 60 segundos."
            status_code: 429
      headers:
        X-RateLimit-Limit:
          description: Límite de requests por ventana
          schema:
            type: integer
        X-RateLimit-Remaining:
          description: Requests restantes
          schema:
            type: integer
        X-RateLimit-Reset:
          description: Segundos hasta reset
          schema:
            type: integer
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API Key de la organización

    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT de autenticación
security:
  - ApiKeyAuth: []
  - BearerAuth: []